remove Value class and subclasses
