+++ runtime/serde/serde-cbor/common/src/aws/smithy/kotlin/runtime/serde/cbor/encoding/Tag.kt
@@ -28,54 +28,17 @@ internal enum class TagId(val value: ULong) {
 }
 
 /**
- * Represents a tagged CBOR [Value] (major type 6). The minor type describes the contents of the tagged value:
- * - 1 -> Timestamp (encoded as epoch seconds)
- * - 2 -> Unsigned bignum
- * - 3 -> Negative bignum
- * - 4 -> Decimal fraction
- */
-internal class Tag(val id: ULong, val value: Value) : Value {
-    override fun encode(into: SdkBufferedSink) {
-        into.writeArgument(Major.TAG, id)
-        value.encode(into)
-    }
-
-    internal companion object {
-        fun decode(buffer: SdkBufferedSource, depth: Int = 0): Tag {
-            val id = decodeArgument(buffer)
-
-            val value: Value = when (id) {
-                TagId.TIMESTAMP.value -> Timestamp.decode(buffer)
-                TagId.BIG_NUM.value -> BigNum.decode(buffer, depth)
-                TagId.NEG_BIG_NUM.value -> NegBigNum.decode(buffer, depth)
-                TagId.DECIMAL_FRACTION.value -> DecimalFraction.decode(buffer, depth)
-                else -> throw DeserializationException("Unsupported tag ID $id")
-            }
-
-            return Tag(id, value)
-        }
-    }
-}
-
-/**
- * Represents a CBOR timestamp, a [Tag] with ID 1.
+ * Write a CBOR timestamp, a tag with ID 1.
  * The tagged value is a number representing the number of seconds since epoch.
- * Note: this number may be an unsigned integer, negative integer, or floating point number.
- * @param value the [Instant] that this CBOR timestamp represents
  */
-internal class Timestamp(val value: Instant) : Value {
-    override fun encode(into: SdkBufferedSink) = into.writeTimestamp(value)
-
-    internal companion object {
-        internal fun decode(buffer: SdkBufferedSource): Timestamp = Timestamp(decodeInstant(buffer))
-    }
-}
-
 internal fun SdkBufferedSink.writeTimestamp(value: Instant) {
     writeArgument(Major.TAG, TagId.TIMESTAMP.value)
     writeFloat64(value.epochMilliseconds / 1000.toDouble())
 }
 
+/**
+ * Decode the value of a CBOR timestamp tag as an [Instant].
+ */
 internal fun decodeInstant(buffer: SdkBufferedSource): Instant {
     val major = peekMajor(buffer)
     val minor = peekMinorByte(buffer)
@@ -103,17 +66,8 @@ internal fun decodeInstant(buffer: SdkBufferedSource): Instant {
 }
 
 /**
- * Represents a CBOR bignum, a [Tag] with ID 2.
- * @param value the [BigInteger] that this CBOR bignum represents.
+ * Write a CBOR bignum, a tag with ID 2.
  */
-internal class BigNum(val value: BigInteger) : Value {
-    override fun encode(into: SdkBufferedSink) = into.writeBigNum(value)
-
-    internal companion object {
-        internal fun decode(buffer: SdkBufferedSource, depth: Int = 0): BigNum = BigNum(decodeBigNum(buffer, depth))
-    }
-}
-
 internal fun SdkBufferedSink.writeBigNum(value: BigInteger) {
     writeArgument(Major.TAG, TagId.BIG_NUM.value)
     writeByteString(value.toByteArray())
@@ -122,17 +76,8 @@ internal fun SdkBufferedSink.writeBigNum(value: BigInteger) {
 internal fun decodeBigNum(buffer: SdkBufferedSource, depth: Int = 0): BigInteger = BigInteger(decodeByteStringValue(buffer, depth))
 
 /**
- * Represents a CBOR negative bignum, a [Tag] with ID 3.
- * @param value the [BigInteger] that this negative CBOR bignum represents.
+ * Write a CBOR negative bignum, a tag with ID 3.
  */
-internal class NegBigNum(val value: BigInteger) : Value {
-    override fun encode(into: SdkBufferedSink) = into.writeNegBigNum(value)
-
-    internal companion object {
-        internal fun decode(buffer: SdkBufferedSource, depth: Int = 0): NegBigNum = NegBigNum(decodeNegBigNum(buffer, depth))
-    }
-}
-
 internal fun SdkBufferedSink.writeNegBigNum(value: BigInteger) {
     val magnitude = NEGATIVE_ONE - value
     val twosComplement = magnitude.toByteArray()
@@ -153,17 +98,8 @@ internal fun decodeNegBigNum(buffer: SdkBufferedSource, depth: Int = 0): BigInte
 }
 
 /**
- * Represents a CBOR decimal fraction, a [Tag] with ID 4.
- * @param value the [BigDecimal] that this decimal fraction represents.
+ * Write a CBOR decimal fraction, a tag with ID 4.
  */
-internal class DecimalFraction(val value: BigDecimal) : Value {
-    override fun encode(into: SdkBufferedSink) = into.writeDecimalFraction(value)
-
-    internal companion object {
-        internal fun decode(buffer: SdkBufferedSource, depth: Int = 0): DecimalFraction = DecimalFraction(decodeDecimalFraction(buffer, depth))
-    }
-}
-
 internal fun SdkBufferedSink.writeDecimalFraction(value: BigDecimal) {
     val mantissaString = value.mantissa.toString()
     val isNegative = mantissaString.startsWith('-')
