+++ runtime/serde/serde-cbor/common/test/aws/smithy/kotlin/runtime/serde/cbor/CborDeserializerTest.kt
@@ -6,7 +6,7 @@ package aws.smithy.kotlin.runtime.serde.cbor
 
 import aws.smithy.kotlin.runtime.io.SdkBuffer
 import aws.smithy.kotlin.runtime.serde.*
-import aws.smithy.kotlin.runtime.serde.cbor.encoding.Value
+import aws.smithy.kotlin.runtime.serde.cbor.encoding.skipValue
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFails
@@ -34,7 +34,7 @@ class CborDeserializerTest {
 
     @Test
     fun testRecursionLimitingThrows() {
-        // Indef map { "x": [0x81 × n, 0x00] } — "x" is unknown → skipValue → Value.decode recursion.
+        // Indef map { "x": [0x81 × n, 0x00] } — "x" is unknown → skipValue recursion.
         val n = MAX_RECURSION_DEPTH + 1
         val p = ByteArray(1 + 2 + n + 1 + 1)
         var i = 0
@@ -64,16 +64,16 @@ class CborDeserializerTest {
         p[n] = 0x00 // uint 0
 
         val buffer = SdkBuffer().apply { write(p) }
-        Value.decode(buffer)
+        skipValue(buffer)
     }
 
     /**
      * tt/P441722592/F9: A CBOR payload with an unknown field containing deeply nested arrays must throw
-     * DeserializationRecursionException, not StackOverflowError. The depth check in Value.decode catches this before
+     * DeserializationRecursionException, not StackOverflowError. The depth check in skipValue catches this before
      * the stack is exhausted.
      */
     @Test
-    fun valueDecodeSkipRecursionThrowsRecursionException() {
+    fun skipValueRecursionThrowsRecursionException() {
         val n = MAX_RECURSION_DEPTH + 1
         val p = ByteArray(1 + 2 + n + 1 + 1)
         var i = 0
@@ -153,29 +153,26 @@ class CborDeserializerTest {
         p[i++] = 0x41 // ByteString(1)
         p[i++] = 0x01 // byte value
 
-        val buffer = SdkBuffer().apply { write(p) }
+        val deserializer = CborPrimitiveDeserializer(SdkBuffer().apply { write(p) })
         // The outermost decimal fraction's mantissa is itself a Tag(4), which is not a valid mantissa → rejected here,
         // before any recursion into the nested levels.
         assertFailsWith<DeserializationException> {
-            Value.decode(buffer)
+            deserializer.deserializeBigDecimal()
         }
     }
 
     /**
-     * Nested indefinite-length TextStrings bypass the depth limit because TextString.decode calls
-     * IndefiniteList.decode(buffer) without a depth parameter (defaults to 0).
-     *
-     * Each indefinite text string level: 0x7F (indef string start) ... 0xFF (break)
-     * Inside, Value.decode dispatches STRING major back to TextString.decode, which again
-     * calls IndefiniteList.decode(buffer) with depth=0, creating unbounded recursion.
+     * Nested indefinite-length text strings are structurally valid CBOR (each level is an indefinite text string whose
+     * only chunk is another indefinite text string), so they are rejected purely on the recursion-depth limit rather
+     * than as malformed input. `decodeTextStringValue` must therefore propagate its depth into each chunk it decodes;
+     * an earlier implementation reset the counter per level and overflowed the stack instead. This exercises the depth
+     * guard on the production `deserializeString` path.
      *
      * Payload: n nested indefinite text strings, innermost contains a definite chunk "a".
      *   0x7F 0x7F 0x7F ... 0x61 0x61 ... 0xFF 0xFF 0xFF
-     *
-     * This should throw DeserializationRecursionException but currently causes StackOverflowError.
      */
     @Test
-    fun indefiniteTextStringResetsDepthCounter() {
+    fun indefiniteTextStringDoesNotBypassDepthLimit() {
         val n = MAX_RECURSION_DEPTH + 1
         // n × 0x7F (indef text start) + definite chunk "a" (0x61 0x61) + n × 0xFF (break)
         val p = ByteArray(n + 2 + n)
@@ -185,9 +182,9 @@ class CborDeserializerTest {
         p[i++] = 0x61 // "a"
         for (j in 0..<n) p[i++] = 0xFF.toByte() // break
 
-        val buffer = SdkBuffer().apply { write(p) }
+        val deserializer = CborPrimitiveDeserializer(SdkBuffer().apply { write(p) })
         assertFailsWith<DeserializationRecursionException> {
-            Value.decode(buffer)
+            deserializer.deserializeString()
         }
     }
